requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyYAML
dill
mlflow
dagshub
fastapi
uvicorn

##-e . 
##Include at the end of the project